Update rts/Sim/Units/UnitHandler.cpp
Crash fix

+++ rts/Sim/Units/UnitHandler.cpp
@@ -81,15 +81,15 @@ CUnitHandler::CUnitHandler()
 		maxUnits += teamHandler->Team(n)->maxUnits;
 	}
 
-	size_t numThreads = std::max(0, configHandler->GetInt("SimThreadCount"));
+	int numThreads = std::max(0, configHandler->GetInt("SimThreadCount"));
 	if (numThreads == 0) {
-		unsigned lcpu = Threading::GetAvailableCores();
-		unsigned pcpu = Threading::GetPhysicalCores();
+		int lcpu = Threading::GetAvailableCores();
+		int pcpu = Threading::GetPhysicalCores();
 		LOG_L(L_WARNING, "CPU: L = %d,  P = %d", lcpu, pcpu);
 		// deduct all logical cores dedicated to the rendering/sim main threads
-		numThreads = std::max((unsigned)GML::NumMainSimThreads(), lcpu - GML::NumMainThreads() * (int)(lcpu / pcpu) + GML::NumMainSimThreads()); // add the main sim thread
+		numThreads = std::max((int)GML::NumMainSimThreads(), lcpu - GML::NumMainThreads() * (int)(lcpu / pcpu) + GML::NumMainSimThreads()); // add the main sim thread
 	}
-	simNumExtraThreads = (!modInfo.multiThreadSim) ? 0 : std::max((size_t)0, numThreads - GML::NumMainSimThreads());
+	simNumExtraThreads = (!modInfo.multiThreadSim) ? 0 : std::max(0, numThreads - GML::NumMainSimThreads());
 	LOG_L(L_WARNING, "CPU: N = %d,  S = %d", numThreads, simNumExtraThreads);
 	Threading::SimThreadCount(simNumExtraThreads + GML::NumMainSimThreads() + (modInfo.asyncPathFinder ? 1 : 0));
